requirements.txt
# 核心依赖
python-dotenv>=0.19.0
pydantic>=1.9.0
aiohttp>=3.8.5

# AI模型接口
openai>=1.0.0

# 工具库
pillow>=9.0.0
requests>=2.31.0

# Redis支持
redis>=4.5.0
pydantic>=1.9.0
pydantic-settings>=2.0.0

# 测试工具
pytest>=6.2.5
pytest-asyncio>=0.15.0
